requirements.txt
line-bot-sdk
uvicorn[standard]
fastapi
openai
supabase
pydub